refactor(live2d): live2d.go -> package live2d
